默认的

+++ src/AcWing243一个简单的整数问题2/Main.java
@@ -57,7 +57,7 @@ public class Main {
           arr[i] = nextInt();
           long b = arr[i]-arr[i-1];
           add(tr1,i,b);
-          add(tr2,i, (long) b *i);
+          add(tr2,i,  b *i);
 
         }
         while (m-->0){
